+++ R/data-sicalis.R
@@ -1,10 +1,12 @@
 #' Spectral curves from three body regions of stripe-tailed yellow finch
 #' (*Sicalis citrina*) males
 #'
-#' @name sicalis
-#' @docType data
+#' Dataset containing reflectance measurements from 3 body parts ("C": crown,
+#' "B": breast, "T": throat) from seven male stripe-tailed yellow finches
+#' (*Sicalis citrina*)
+#'
+#' @usage
+#' data(sicalis)
+#'
 #' @author Rafael Maia \email{rm72@@zips.uakron.edu}
-#' @keywords data
-#' @description dataset containing reflectance measurements from 3 body parts
-#' ("C": crown, "B": breast, "T": throat) from seven male stripe-tailed yellow finches
-NULL
+"sicalis"