+++ R/summary.colspace.R
@@ -61,6 +61,21 @@ summary.colspace <- function(object, by = NULL, ...) {
     return(summary(as.data.frame(object)))
   }
 
+
+  if (!is.null(attr(object, "data.maxgamut"))) {
+    maxgamut <- attr(object, "data.maxgamut")
+    if (attr(object, "clrsp") == "dispace") {
+      maxvol <- max(maxgamut) - min(maxgamut)
+    } else {
+      maxvol <- tryCatch(
+        convhulln(attr(object, "data.maxgamut"), "FA")$vol,
+        error = function(e) NA
+      )
+    }
+  } else {
+    maxvol <- NA
+  }
+
   cat(
     "Colorspace & visual model options:\n",
     "* Colorspace:", attr(object, "clrsp"), "\n",
@@ -69,10 +84,16 @@ summary.colspace <- function(object, by = NULL, ...) {
     "* Visual system, achromatic:", attr(object, "visualsystem.achromatic"), "\n",
     "* Illuminant:", attr(object, "illuminant"), "\n",
     "* Background:", attr(object, "background"), "\n",
-    "* Relative:", attr(object, "relative"), "\n", "\n"
+    "* Relative:", attr(object, "relative"), "\n",
+    "* Max possible chromatic volume:", maxvol, "\n"
   )
 
-  if (attr(object, "clrsp") != "tcs") summary.data.frame(object)
+
+  cat("\n")
+
+  if (attr(object, "clrsp") != "tcs") {
+    return(summary.data.frame(object))
+  }
 
   if (attr(object, "clrsp") == "tcs") {
     if (!is.null(by)) {