+++ tests/testthat/test-colspace.R
@@ -150,8 +150,8 @@ test_that("Errors/messages", {
   expect_warning(colspace(vismodel(flowers, visual = sensmodel(c(300, 400, 500, 600, 700))), space = "tcs"), "not tetrachromatic")
   class(vis.flowers) <- "data.frame"
   expect_error(colspace(vis.flowers[1:3], space = "tcs"), "has fewer than four")
-  expect_warning(colspace(vis.flowers, space = "tcs"), "treating columns as")
-  expect_warning(colspace(cbind(vis.flowers, vis.flowers[1:2]), space = "tcs"), "has more than four columns")
+  expect_message(colspace(vis.flowers, space = "tcs"), "treating columns as")
+  expect_message(colspace(cbind(vis.flowers, vis.flowers[1:2]), space = "tcs"), "has more than four columns")
 
   vis.flowers <- vismodel(flowers, visual = "bluetit")
   names(vis.flowers) <- c("a", "b", "c", "d", "e")
@@ -159,4 +159,15 @@ test_that("Errors/messages", {
 
   expect_equivalent(round(sum(summary(colspace(vismodel(flowers)))), 5), 4.08984)
   expect_equivalent(round(sum(summary(colspace(vismodel(flowers))), by = 3), 5), 7.08984)
+})
+
+test_that("CIE", {
+  
+  data(flowers)
+  vis_flowers <- colspace(vismodel(flowers, "cie10"))
+  implicit_cie_flowers <- colspace(vis_flowers)
+  expect_s3_class(implicit_cie_flowers, "colspace")
+  explicit_cie_flowers <- colspace(vis_flowers, space = "ciexyz")
+  expect_identical(implicit_cie_flowers, explicit_cie_flowers)
+  
 })